pyproject.toml
[project]
name = "omicvision"
version = "0.0.1"
description = ""
readme = "README.md"
authors = [
    { name = "Lars Andersen Bratholm", email = "larsbratholm@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic",
    "jaxtyping",
    "numpy",
    "cvxpy[ecos]",
    "transformers",
    "seaborn",
    "pyscipopt",
    "torch",
    "accelerate",
    "bitsandbytes",
    "openpyxl>=3.1.5",
    "scikit-learn>=1.6.1",
]

[dependency-groups]
dev = ["pre-commit", "mypy", "pandas-stubs"]


[tool.ruff.lint]
# Add docstyle convensions
select = ["D", "I", "E4", "E7", "E9", "F"]
extend-safe-fixes = ["D415"]
ignore = [
"D200", # Disable condensing single-line docstrings
"D212", # Start docstring summary on second line (enables D213)
"D107", # Ignore init docstrings
"F722", # jaxtyping compability
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
# Enable reformatting of code snippets in docstrings.
docstring-code-format = true

[tool.mypy]
strict = true
implicit_reexport = true

[[tool.mypy.overrides]]
module = ["transformers"]
ignore_missing_imports = true
